Fix building on Fedora 40

+++ src/channelhopper.c
@@ -45,7 +45,7 @@ int chans_in [MAX_CHAN_COUNT] = {0};
 int chans_out [MAX_CHAN_COUNT] = {0};
 
 pthread_t *hopper = NULL;
-pthread_t chan_sniffer = NULL;
+pthread_t chan_sniffer = (int)NULL;
 int hopper_useconds = 0;
 volatile int sniff = 0;
 pthread_mutex_t chan_thread_mutex;
